+++ common/hybriddb.cpp
@@ -60,10 +60,10 @@ bool hybriddb::getParams( int code, dbrecord &recout )
 		return false;
 	}
 
-	if( (recout.in_start = getintparam( bufptr, false, true )) == - 1
-			|| (recout.in_end = getintparam( bufptr, false, true )) ==  -1
-			|| (recout.out_start = getintparam( bufptr, false, true )) ==  -1
-			|| (recout.out_end = getintparam( bufptr, false, true )) ==  -1
+	if( (recout.in_start = getintparam( bufptr, false, true )) == (uint16_t) - 1
+			|| (recout.in_end = getintparam( bufptr, false, true )) == (uint16_t) -1
+			|| (recout.out_start = getintparam( bufptr, false, true )) == (uint16_t) -1
+			|| (recout.out_end = getintparam( bufptr, false, true )) == (uint16_t) -1
 			|| (tmp = getintparam( bufptr, false, true )) ==  -1 )
 	{
 		recout.in_start = -1;
